Extract media sidebar state hook and utils from agents-view.tsx (1034 → 925 lines)
